cleanup `client` package (#66)
* small nits

* nit

* nit

* cleanup

* nits

* regen mocks

* nit

* rename

* fix spelling

* reuse eth client

* nits

* remove rpc file

* nits

* nit

* fix

* fully populate

* revert

* nit

* fix merge error

* nit

* cleanup

+++ service/service_account.go
@@ -101,7 +101,7 @@ func (s AccountService) AccountBalance(
 			return nil, wrapError(errInternalError, err)
 		}
 
-		amount := mapper.Erc20Amount(response, contractAddress, currency.Symbol, uint8(currency.Decimals), false)
+		amount := mapper.Erc20Amount(response, contractAddress, currency.Symbol, currency.Decimals, false)
 
 		balances = append(balances, amount)
 	}
